refactor: changes on clean arch
The layers of the app were not well implemented, so decided to refactor for a properly architechture design.
Now the layers are decoupled from each other and only being dependants based on clean architechture.
Also implemented FileStorage service for a future cloud static files storage and management also to be used on local